// src/app/core/enums/fragments.ts
export enum HomeFragment {
  home = 'home-section',
  services = 'services-section',
  whyUs = 'why-us-section',
  steps = 'steps-section',
  pricing = 'cost-breakdown-section',
  ourWorks = 'our-works-section',
  contactUs = 'contact-us-section',
  testimonials = 'testimonials-section',
}

export enum AboutFragment {
  about = 'about-section',
  whoWeAre = 'who-we-are-section',
  clientele = 'clientele-section',
  brochure = 'brochure-section',
  bosses = 'bosses-section',
  staff = 'staff-section',
  joinUs = 'join-us-section',
}

export enum ArticlesFragment {
  blog = 'blog-section',
  industryNews = 'industry-news-section',
}

export enum BusinessFragment {
  business = 'business-section',
  grow = 'grow-business-section',
  banner = 'banner-section',
  advantages = 'advantages-section',
  getInTouch = 'get-in-touch-section',
}

export enum CareersFragment {
  careers = 'careers',
  benefits = 'perks-and-benefits',
  openPositions = 'open-positions',
}

export enum ApplyCareerFragment {
  form = 'application-form-section',
}

export enum FooterFragment {
  footer = 'footer-section',
}