// src/app/core/enums/routers.ts
export enum Routers {
  Login = '/login',
  Register = '/get-access',
  ForgotPassword = '/forgot-password',
  Home = '/home',
  About = '/about',
  Articles = '/articles',
  Business = '/business',
  Contact = '/contact',
  Careers = '/careers',
  Accounts = '/accounts',
  Position = '/position',
  Application = '/apply',
  LegalNotes = '/legal-notes',
  Privacy = '/legal-notes/privacy-policy',
  Terms = '/terms-conditions',
  EmployeeForms = '/employee-forms',
  SuccessSubmission = '/employee-forms/success',
  Approved = '/approved',
  Declined = '/declined',
  EmployeeFormsLogin = '/emp-forms-login',
  EmployeeFormsRegister = '/emp-forms-register',
}

export enum TRouters {
  Login = 'login',
  Register = 'get-access',
  ForgotPassword = 'forgot-password',
  Home = 'home',
  About = 'about',
  Articles = 'articles',
  Business = 'business',
  Contact = 'contact',
  Careers = 'careers',
  Accounts = 'accounts',
  Position = 'position/:id',
  Application = 'apply/:id',
  LegalNotes = 'legal-notes',
  Privacy = 'privacy-policy',
  Terms = 'terms-conditions',
  Forms = 'employee-forms',
  Success = 'success',
  Approved = 'approved',
  Declined = 'declined',
  EmployeeFormsLogin = 'emp-forms-login',
  EmployeeFormsRegister = 'emp-forms-register',
}